+++ assember_test.cpp
@@ -2,7 +2,7 @@
 #pragma ide diagnostic ignored "cert-err58-cpp"
 
 #include <gtest/gtest.h>
-#include "myerror.h"
+#include "syntaxanalyzer.h"
 #include "fileworker.h"
 
 // TODO: add stack tests overflow
